% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{depth}
\alias{depth}
\title{Compute the univariate tukey depth of a function \code{g} with respect to a set of functions \code{fmat}}
\usage{
depth(g, fmat)
}
\arguments{
\item{g}{Vector. Function that you want to find the univariate tukey depths of.}

\item{fmat}{Matrix of functions. Each column is a function. Need to be same length as g.}
}
\value{
Vector of pointwise tukey depths
}
\description{
Compute the univariate tukey depth of a function \code{g} with respect to a set of functions \code{fmat}
}
