% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gp1d}
\alias{gp1d}
\title{Generates realizations of one dimensional Gaussian process with exponential covariance function}
\usage{
gp1d(n = 100, mu = 0, sd = 1, r = 10, pts = 50)
}
\arguments{
\item{n}{Positive integer. Number of functions to generate.}

\item{mu}{numeric value or vector of length \code{pts} Mean value of the Gaussian process. Defaults to 0.}

\item{sd}{Positive numeric or vector of length \code{pts} Standard deviation of the Gaussian process. Defaults to 1.}

\item{r}{Positive numeric or vector of length \code{pts} Range of the Gaussian process. Defaults to 10.}

\item{pts}{Positive integer. Grid size to sample each functions on. Defaults to 50.}
}
\value{
Matrix of functions. Each column is a separate function.
}
\description{
Generates realizations of one dimensional Gaussian process with exponential covariance function
}
\note{
Setting a single number in \code{mu}, \code{sd}, or \code{r} means that the parameter will be constant over all observation
points. Setting it to a vector allows you to vary it by observation site. Vector length must be equal to \code{pts}.
}
