% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gp2d}
\alias{gp2d}
\title{Generates realizations of two dimensional Gaussian process with exponential covariance function}
\usage{
gp2d(n = 100, mu = 0, sd = 1, l = 30, pts = 30)
}
\arguments{
\item{n}{Positive integer. Number of functions to generate.}

\item{mu}{numeric value or matrix of size \code{pts^2} Mean value of the Gaussian process. Defaults to 0.}

\item{sd}{Positive numeric or vector of length \code{pts^2} Standard deviation of the Gaussian process. Defaults to 1.}

\item{pts}{Positive integer. Functions will be sampled on a \code{pts} x \code{pts} grid. Defaults to 30.}

\item{r}{Positive numeric or vector of length \code{pts^2} Range of the Gaussian process. Defaults to 30.}
}
\value{
3D array of functions. First two indicies correspond to values in the 2D GP. Third index is the function number.
}
\description{
Generates realizations of two dimensional Gaussian process with exponential covariance function
}
\note{
Setting a single number in \code{mu}, \code{sd}, or \code{r} means that the parameter will be constant over all observation
points. Setting it to a vector allows you to vary it by observation site. Vector length must be equal to \code{pts}.
}
