% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{int_depth}
\alias{int_depth}
\title{Compute the integrated tukey depth of each function in  \code{gmat} with respect to a set of functions \code{fmat}}
\usage{
int_depth(gmat, fmat)
}
\arguments{
\item{gmat}{Matrix of functions. Each column is a function.}

\item{fmat}{Matrix of functions. Each column is a function. Need to be same length as gmat.}
}
\value{
Vector of integrated tukey depths
}
\description{
Compute the integrated tukey depth of each function in  \code{gmat} with respect to a set of functions \code{fmat}
}
