% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{ks_cdf}
\alias{ks_cdf}
\title{Calculates the probability of the value \code{x} under the Kolmogorov Distribution.}
\usage{
ks_cdf(x, n = 20)
}
\arguments{
\item{x}{A positive number.}

\item{n}{A positive number. Number of terms to include in the Kolmogorov distribution. Defaults to 20.}
}
\value{
Probability of x.
}
\description{
Calculates the probability of the value \code{x} under the Kolmogorov Distribution.
}
