% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{kstat}
\alias{kstat}
\title{Tests if the functions in \code{gmat} have the same distribution as the functions in \code{fmat}}
\usage{
kstat(fmat, gmat)
}
\arguments{
\item{fmat}{Matrix of functions. Each column is a function. Need to be same length as gmat.}

\item{gmat}{Matrix of functions. Each column is a function.}
}
\value{
Value of the K statistic and the associated p value under the null, as a vector.
}
\description{
Tests if the functions in \code{gmat} have the same distribution as the functions in \code{fmat}
}
